Revert "Make build reproducible by forcing an empty GNU build-id"
This reverts commit 1460cfa128fb457b5b5c60fcc5cac6faf5a216d5.

After discussion on the reproducible-builds@lists.alioth.debian.org
it was decided this fix is wrong, as it most likely makes the debug
symbols unusable, and fixed a non-issue as the same old revision
(before this change, that was never uploaded) started to work on
reproducible-builds.org testing infra. So the root cause of the
failure apparently was something in the r-b.o infra, not this Galera.