Follow upstream change and switch from SCons to CMake
Upstream changed the build system for Galera 3 from SCons to
CMake in version 25.3.32. Old build stopped working, so we must
follow the upstream change and also use CMake from now on.

The debian/rules code is the same as in src:galera-4 so it should
be well tested and the optimal solution for Galera 3 as well.

Also adapt *.install and *.links to upstream build system changes.
The files are now automatically in the correct location and the
Debian packaging has less move around to do.